backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
cloudinary==1.36.0
reportlab==4.0.9
